Two New Ways to Earn Money
When the 5th item, the star-dust tree, is built, it can be clicked on
the main scene to earn a small amount of money. When the 6th item, the
galactic stone, is built, it can be clicked on the main scene to pull up
a prompt. The user can then choose whether or not to touch the stone,
which results in either making a moderate amount of money or losing a
lot of money. It's a gamble.